Fixed buildClientRetainImg
Client build file and associated gradle tasks should all work correctly now.

Currently it compiles all code in the Client module, but it points at the ClientTest class for the main method. This can be changed by modifying the Dockerfile in the client root directory, and replacing any occurences of ClientTest in the Client Gradle Build file with whatever  Main Class you want to use instead