Improved the Test module
Added tasks to create a popup terminal window for each operating system (use gradlew tasks to find the appropriate task for your OS).

Added a TestManager class and TestRunner class.

TestManager is the class you can add your method calls to. Ideally you should create a Class in the appropriate tests directory with a static method and then call it in TestsManager.