Proper error handling and closing of channels implemented.
If a channel triggers a read event on the host process, but then the remote process closes it, this can result in an error if the host tries to continue reading from that channel, or tries to write to it in the future.

Proper error checking and handling has been implemented to avoid this - if a SocketChannel is closed by the remote connection, the server recognizes it, removes it's key from the selector, closes the NIOMessageChannel (removes it from the Hashmap), and closes the SocketChannel. It then continues the main event loop.