Overhauled the handling of connection closed and IO exceptions
ConnectionCleanupManager is now responsible for maintaining a conistent set of communication channels and server records.

AddressingServer, BroadcastManager and NetworkManager all keep an instance of this class so that errors can be caught and dispatched appropriately.

BroadcastManager sends a single message accross many channels, so it needs to be able to cleanup any bad connections immediately before it returns control to the calling code.