Added retry logic methods to PendingEvent
PendingEvents now have iteration counts, and can easily be used to try resending a message to any processes that haven't responded.

Additionally, I added proper handling to the broadcast manager so that it removes any channels and PIDs from the pending event if a message fails to send (i.e. we don't wait for an ACK from a process that never got the message)