Added retry logic to PendingEvent
All PendingEvents (used for SC) now have retry logic - i.e. update messages are resent twice.

If an update message is failed to receive an ACK after 3 attempts the process(es) that failed to ACK are removed from the network and the requester is sent a NotificationMessage with ObjectType.REQUEST_FAILURE

Also centralized replication synching to ReplicaSyncCoordinator

Created the PendingEventWatchdog class for thread use.

Added proper closure of thread to AddrServerNetworkManager.

Ensured all methods in PeningEventWatchdog are thread-safe and allow for concurrent access of shared resources.

Added new NotifcationMessage factory method. This message is sent to any process that made a request which has failed to be processed by the PRIMARY addressing server.